Fix Windows build (#396)
* cast numel to long

* remove comment

+++ autogptq_extension/exllama/exllama_ext.cpp
@@ -116,13 +116,13 @@ void prepare_buffers
     int device_index = device.index();
     TORCH_CHECK_DEVICE_INDEX(device_index);
     const at::cuda::OptionalCUDAGuard device_guard(device);
-    const long max_int = std::numeric_limits<int>::max();  // temp_state.numel() is long
+    const long max_int = std::numeric_limits<int>::max();
 
     prepare_buffers_cuda
     (
         device_index,
         // buffer size used for sanity checks
-        std::clamp(temp_state.numel(), (long)0, max_int),
+        std::clamp((long)temp_state.numel(), (long)0, max_int),
         (half*) temp_state.data_ptr(),
         (half*) temp_dq.data_ptr()
     );
